refactor(user-management): B2BCAT-40 remove unused code from checkUserEmail
